DynamicCast: avoid compile-time errors re "ambiguous cast"
Imagine the following:

* class A
* class B : A
* class C : A
* class D : B, C

We can't static_cast<A*>(D*), because we have 2 instances of A in D.
The list of all available casts is generated during the compilation,
but currently there's no check whether the cast can be performed.
Use C++20's concept to handle this.
Note that dynamic_cast<A*>(D*) would behave the same -- ambiguous
casts are not allowed.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>